Top-k selection and partitioning on MLX tensors

Top-k selection and partitioning on MLX tensors

Usage

mlx_topk(x, k, axis = NULL)

mlx_partition(x, kth, axis = NULL)

mlx_argpartition(x, kth, axis = NULL)

Arguments

x

An object coercible to mlx.

k

Positive integer specifying the number of elements to select.

axis

Optional axis to operate over (1-indexed like R). When NULL, the tensor is flattened first.

kth

Zero-based index of the element that should be placed in-order after partitioning.

Value

An mlx tensor.

Details

mlx_topk() returns the largest k values as reported by MLX. Use mlx_argsort() if you need the associated indices.

See also

Examples

scores <- as_mlx(c(0.7, 0.2, 0.9, 0.4))
mlx_topk(scores, k = 2)
#> mlx array [2]
#>   dtype: float32
#>   device: gpu
#>   values:
#> [1] 0.7 0.9
mlx_partition(scores, kth = 1)
#> mlx array [4]
#>   dtype: float32
#>   device: gpu
#>   values:
#> [1] 0.2 0.4 0.7 0.9
mlx_argpartition(scores, kth = 1)
#> mlx array [4]
#>   dtype: uint32
#>   device: gpu
#>   values:
#> [1] 1 3 0 2
mlx_topk(as_mlx(matrix(1:6, 2, 3)), k = 1, axis = 1)
#> mlx array [1 x 3]
#>   dtype: float32
#>   device: gpu
#>   values:
#>      [,1] [,2] [,3]
#> [1,]    2    4    6